requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2.4
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.93.0
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
